+++ packages/api-headless-cms/__tests__/utils/helpers.ts
@@ -1,5 +1,7 @@
 import { pick } from "lodash";
 import { SecurityIdentity } from "@webiny/api-security/types";
+export { until } from "@webiny/project-utils/testing/helpers/until";
+export { sleep } from "@webiny/project-utils/testing/helpers/sleep";
 
 export interface PermissionsArg {
     name: string;
@@ -75,54 +77,4 @@ export const getModelCreateInputObject = model => {
         "lockedFields",
         "titleFieldId"
     ]);
-};
-
-type UntilOptions = {
-    name?: string;
-    tries?: number;
-    wait?: number;
-};
-
-export const sleep = (ms = 333) => {
-    return new Promise(resolve => {
-        setTimeout(resolve, ms);
-    });
-};
-
-export const until = async (execute, until, options: UntilOptions = {}) => {
-    const { name = "NO_NAME", tries = 5, wait = 300 } = options;
-
-    let result;
-    let triesCount = 0;
-
-    while (true) {
-        result = await execute();
-
-        let done;
-        try {
-            done = await until(result);
-        } catch {}
-
-        if (done) {
-            return result;
-        }
-
-        triesCount++;
-        if (triesCount === tries) {
-            break;
-        }
-
-        // Wait.
-        await new Promise((resolve: any) => {
-            setTimeout(() => resolve(), wait);
-        });
-    }
-
-    throw new Error(
-        `[${name}] Tried ${tries} times but failed. Last result that was received: ${JSON.stringify(
-            result,
-            null,
-            2
-        )}`
-    );
 };