+++ packages/app-page-builder/src/admin/graphql/pageImportExport.gql.ts
@@ -18,9 +18,17 @@ stats {
 }`;
 
 export const IMPORT_PAGES = gql`
-    mutation PbImportPage($category: String!, $data: PbImportPageInput!) {
+    mutation PbImportPage(
+        $category: String!,
+        $zipFileKey: String,
+        $zipFileUrl: String
+    ) {
         pageBuilder {
-            importPages(category: $category, data: $data) {
+            importPages(
+                category: $category,
+                zipFileKey: $zipFileKey,
+                zipFileUrl: $zipFileUrl
+            ) {
                 data {
                     task {
                         id
@@ -36,9 +44,21 @@ export const IMPORT_PAGES = gql`
 `;
 
 export const EXPORT_PAGES = gql`
-    mutation PbExportPages($ids: [ID]!, $revisionType: PbExportPageRevisionType) {
+    mutation PbExportPages(
+        $ids: [ID!],
+        $revisionType: PbExportPageRevisionType!,
+        $where: PbListPagesWhereInput,
+        $sort: [PbListPagesSort!],
+        $search: PbListPagesSearchInput
+    ) {
         pageBuilder {
-            exportPages(ids: $ids, revisionType: $revisionType) {
+            exportPages(
+                ids: $ids,
+                revisionType: $revisionType,
+                where: $where,
+                sort: $sort,
+                search: $search
+            ) {
                 data {
                     task {
                         id