+++ packages/app-page-builder/src/editor/recoil/actions/updateElement/types.ts
@@ -1,6 +1,7 @@
 import { PbEditorElement } from "~/types";
+import { SaveRevisionActionArgsType } from "~/editor/recoil/actions/saveRevision/types";
 
-export type UpdateElementActionArgsType = {
+export type UpdateElementActionArgsType = SaveRevisionActionArgsType & {
     element: PbEditorElement;
     history: boolean;
     triggerUpdateElementTree?: boolean;