+++ packages/app-page-builder/CHANGELOG.md
@@ -3,6 +3,44 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [5.17.0](https://github.com/webiny/webiny-js/compare/v5.17.0-beta.2...v5.17.0) (2021-11-08)
+
+**Note:** Version bump only for package @webiny/app-page-builder
+
+
+
+
+
+# [5.17.0-beta.2](https://github.com/webiny/webiny-js/compare/v5.17.0-beta.1...v5.17.0-beta.2) (2021-11-08)
+
+
+### Bug Fixes
+
+* **app-page-builder:** fix menu styles ([9dbcf62](https://github.com/webiny/webiny-js/commit/9dbcf62c5c2d598ff281ccee980bfd99c36fa772))
+
+
+
+
+
+# [5.17.0-beta.1](https://github.com/webiny/webiny-js/compare/v5.16.0...v5.17.0-beta.1) (2021-11-08)
+
+
+### Bug Fixes
+
+* **app-page-builder:** fix Broken sort settings field for List of Pages ([#1997](https://github.com/webiny/webiny-js/issues/1997)) ([61225b3](https://github.com/webiny/webiny-js/commit/61225b3c21de0ecf7439372767a0077f87799182)), closes [#1994](https://github.com/webiny/webiny-js/issues/1994)
+* **app-page-builder:** fix incorrect width applied to Grid element inside editor ([#2007](https://github.com/webiny/webiny-js/issues/2007)) ([f87635e](https://github.com/webiny/webiny-js/commit/f87635e744866107c3a6b86e7b7048531dc77b7f))
+* **app-page-builder:** use debounce in updateElementAction inside Text element ([b09fb30](https://github.com/webiny/webiny-js/commit/b09fb30ecc916ec4eeb02bee85ef03d57b624db6))
+
+
+### Features
+
+* **app-page-builder:** add debounce in updateElementAction params ([86b4460](https://github.com/webiny/webiny-js/commit/86b446072f1d375e44f8085e520fff8ab2b54ec0))
+* new security and multi-tenancy ([#1993](https://github.com/webiny/webiny-js/issues/1993)) ([99ecfca](https://github.com/webiny/webiny-js/commit/99ecfca45011cb0f6476c3bc5c883c56a055aaa0))
+
+
+
+
+
 # [5.16.0](https://github.com/webiny/webiny-js/compare/v5.16.0-beta.4...v5.16.0) (2021-10-21)
 
 **Note:** Version bump only for package @webiny/app-page-builder